# Backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/manufacturers/profile/distributors/{distributorId}": {
      "post": {
        "operationId": "ManufacturerController_linkDistributor_v1",
        "parameters": [
          {
            "name": "distributorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Link Distributor",
        "tags": [
          "Manufacturer"
        ]
      }
    },
    "/v1/distributors/profile": {
      "get": {
        "operationId": "DistributorController_getProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "Distributor"
        ]
      },
      "put": {
        "operationId": "DistributorController_updateProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDistributorProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "Distributor"
        ]
      }
    },
    "/v1/distributors": {
      "get": {
        "operationId": "DistributorController_getDistributors_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Distributor"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Distributors",
        "tags": [
          "Distributor"
        ]
      },
      "post": {
        "operationId": "DistributorController_createDistributorAdmin_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDistributorAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Distributor Admin",
        "tags": [
          "Distributor"
        ]
      }
    },
    "/v1/distributors/{id}": {
      "get": {
        "operationId": "DistributorController_getDistributorById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Distributor By Id",
        "tags": [
          "Distributor"
        ]
      },
      "patch": {
        "operationId": "DistributorController_updateDistributorAdmin_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDistributorAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Distributor Admin",
        "tags": [
          "Distributor"
        ]
      },
      "delete": {
        "operationId": "DistributorController_deleteDistributorAdmin_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete Distributor",
        "tags": [
          "Distributor"
        ]
      }
    },
...
    "/v1/backorders": {
      "get": {
        "operationId": "BackordersController_listBackorders_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Backorder"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Backorders",
        "tags": [
          "Backorders"
        ]
      }
    },
    "/v1/backorders/{id}": {
      "get": {
        "operationId": "BackordersController_getBackorder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorder",
        "tags": [
          "Backorders"
        ]
      }
    },
    "/v1/backorders/{id}/allocate": {
      "post": {
        "operationId": "BackordersController_allocateBackorder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Allocate Backorder",
        "tags": [
          "Backorders"
        ]
      }
    },
...
    "/v1/orders/backorders": {
      "get": {
        "operationId": "OrdersController_getBackorders_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "SortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "SortOrder",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "StartDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "EndDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "distributor_id",
            "required": false,
            "in": "query",
            "description": "Filter by specific distributor UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesman_id",
            "required": false,
            "in": "query",
            "description": "Filter by specific salesman UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorders (paginated)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/orders/backorders/{id}": {
      "get": {
        "operationId": "OrdersController_getBackorderById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Backorder by ID",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Order By Id (includes line items)",
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_updateOrder_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit Order (SALESMAN/DISTRIBUTOR_ADMIN/MANUFACTURER_ADMIN)",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vikas Inventory RN Platform API",
    "description": "Enterprise Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "UpdateDistributorProfileDto": {
        "type": "object",
        "properties": {
          "business_name": {
            "type": "string",
            "description": "Business name"
          },
          "owner_name": {
            "type": "string",
            "description": "Owner name"
          },
          "phone": {
            "type": "string",
            "description": "Phone"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "gst_number": {
            "type": "string",
            "description": "Gst number"
          },
          "address": {
            "type": "string",
            "description": "Address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "pincode": {
            "type": "string",
            "description": "Pincode"
          },
          "distributor_discount_percent": {
            "type": "number",
            "description": "Distributor discount percent"
          }
        }
      },
      "Distributor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id"
          },
          "user_id": {
            "type": "string",
            "description": "User id"
          },
          "business_name": {
            "type": "string",
            "description": "Business name"
          },
          "owner_name": {
            "type": "string",
            "description": "Owner name"
          },
          "phone": {
            "type": "string",
            "description": "Phone"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "gst_number": {
            "type": "string",
            "description": "Gst number"
          },
          "address": {
            "type": "string",
            "description": "Address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "pincode": {
            "type": "string",
            "description": "Pincode"
          },
          "distributor_discount_percent": {
            "type": "number",
            "description": "Distributor discount percent"
          },
          "approval_status": {
            "type": "string",
            "description": "Approval status"
          },
          "approved_by_user_id": {
            "type": "string",
            "description": "Approved by user id"
          },
          "approved_at": {
            "format": "date-time",
            "type": "string",
            "description": "Approved at"
          },
          "rejected_reason": {
            "type": "string",
            "description": "Rejected reason"
          },
          "is_active": {
            "type": "boolean",
            "description": "Is active"
          },
          "is_internal_distributor": {
            "type": "boolean",
            "description": "Is internal distributor"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at"
          },
          "deleted_at": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at"
          }
        },
        "required": [
          "id",
          "business_name",
          "country",
          "distributor_discount_percent",
          "approval_status",
          "is_active",
          "is_internal_distributor",
          "created_at",
          "updated_at",
          "deleted_at"
        ]
      },
      "CreateDistributorAdminDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Password"
          },
          "manufacturer_ids": {
            "description": "Manufacturer ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "business_name": {
            "type": "string",
            "description": "Business name"
          },
          "contact_person": {
            "type": "string",
            "description": "Contact person"
          },
          "phone": {
            "type": "string",
            "description": "Phone"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "gst_number": {
            "type": "string",
            "description": "Gst number"
          },
          "is_active": {
            "type": "boolean",
            "description": "Is active"
          },
          "is_internal_distributor": {
            "type": "boolean",
            "description": "Is internal distributor"
          },
          "address": {
            "type": "string",
            "description": "Address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "pincode": {
            "type": "string",
            "description": "Pincode"
          },
          "distributor_discount_percent": {
            "type": "number",
            "description": "Distributor discount percent"
          }
        },
        "required": [
          "password",
          "manufacturer_ids",
          "business_name",
          "phone",
          "email"
        ]
      },
      "UpdateDistributorAdminDto": {
        "type": "object",
        "properties": {
          "business_name": {
            "type": "string",
            "description": "Business name"
          },
          "owner_name": {
            "type": "string",
            "description": "Owner name"
          },
          "phone": {
            "type": "string",
            "description": "Phone"
          },
          "email": {
            "type": "string",
            "description": "Email"
          },
          "gst_number": {
            "type": "string",
            "description": "Gst number"
          },
          "address": {
            "type": "string",
            "description": "Address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "state": {
            "type": "string",
            "description": "State"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "pincode": {
            "type": "string",
            "description": "Pincode"
          },
          "distributor_discount_percent": {
            "type": "number",
            "description": "Distributor discount percent"
          },
          "is_active": {
            "type": "boolean",
            "description": "Is active"
          },
          "manufacturer_ids": {
            "description": "Manufacturer ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Backorder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id"
          },
          "order_id": {
            "type": "string",
            "description": "Order id"
          },
          "order_item_id": {
            "type": "string",
            "description": "Order item id"
          },
          "product_id": {
            "type": "string",
            "description": "Product id"
          },
          "distributor_id": {
            "type": "string",
            "description": "Distributor id"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity"
          },
          "status": {
            "type": "string",
            "description": "Status"
          },
          "resolved_quantity": {
            "type": "number",
            "description": "Resolved quantity"
          },
          "resolved_at": {
            "format": "date-time",
            "type": "string",
            "description": "Resolved at"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Created at"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at"
          }
        },
        "required": [
          "id",
          "order_id",
          "product_id",
          "distributor_id",
          "quantity",
          "status",
          "resolved_quantity",
          "created_at",
          "updated_at"
        ]
      },
...
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "products": {
            "description": "Products",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string",
            "description": "Reason"
          },
          "manufacturerId": {
            "type": "string",
            "description": "Manufacturer ID (for Draft Distributor->Manufacturer orders)"
          },
          "transportMode": {
            "type": "string",
            "description": "Transport Mode"
          },
          "standardDiscountPercent": {
            "type": "number",
            "description": "Standard Discount Percent"
          },
          "specialDiscountPercent": {
            "type": "number",
            "description": "Special Discount Percent"
          },
          "distributorDiscountPercent": {
            "type": "number",
            "description": "Distributor Discount Percent"
          },
          "distributorMarginPercent": {
            "type": "number",
            "description": "Distributor Margin Percent"
          },
          "freightDiscountPercent": {
            "type": "number",
            "description": "Freight Discount Percent"
          },
          "cashDiscountPercent": {
            "type": "number",
            "description": "Cash Discount Percent"
          }
        },
        "required": [
          "products"
        ]
      },
...
    }
  }
}